Added Mail Server support (TestContainer), better logging support, cleanup
- note: depends on PR for the TCK, but can run even without that:
  https://github.com/eclipse-ee4j/jakartaee-tck/pull/1115
