+++ appserver/tests/tck/tck-runner/src/main/java/org/glassfish/main/tests/tck/ant/TckConfiguration.java
@@ -23,48 +23,78 @@ import java.io.InputStream;
 import java.nio.file.Path;
 import java.util.Properties;
 
+
 /**
+ * Configuration of the TCK runnable by this project.
+ *
  * @author David Matejcek
  */
 public class TckConfiguration {
 
     private final Properties cfg;
 
+
+    /**
+     * Loads the configuration file.
+     *
+     * @param cfgFile
+     */
     public TckConfiguration(File cfgFile) {
         this.cfg = new Properties();
         try (FileInputStream is = new FileInputStream(cfgFile)) {
             this.cfg.load(is);
         } catch (IOException e) {
-            throw new IllegalStateException("Could not load TCK configuration from '" + cfgFile + "'!");
+            throw new IllegalStateException("Could not load the TCK configuration from '" + cfgFile + "'!");
         }
     }
 
 
+    /**
+     * Loads the configuration.
+     *
+     * @param cfgStream
+     */
     public TckConfiguration(InputStream cfgStream) {
         this.cfg = new Properties();
         try {
-            this.cfg.load(cfgStream);
+            try {
+                this.cfg.load(cfgStream);
+            } finally {
+                cfgStream.close();
+            }
         } catch (IOException e) {
-            throw new IllegalStateException("Could not load TCK configuration!");
+            throw new IllegalStateException("Could not load the TCK configuration!");
         }
     }
 
 
+    /**
+     * @return version of the TCK ZIP artifact.
+     */
     public String getTckVersion() {
         return cfg.getProperty("tck.version");
     }
 
 
+    /**
+     * @return .../target/jakartaeetck
+     */
     public File getJakartaeeDir() {
         return new File(getTargetDir(), "jakartaeetck");
     }
 
 
+    /**
+     * @return .../target/jakartaeetck/docker/run_jakartaeetck.sh
+     */
     public Path getJakartaeetckCommand() {
         return getJakartaeeDir().toPath().resolve(Path.of("docker", "run_jakartaeetck.sh"));
     }
 
 
+    /**
+     * @return .../target
+     */
     public File getTargetDir() {
         String dir = cfg.getProperty("target.directory");
         if (dir == null) {
@@ -74,6 +104,9 @@ public class TckConfiguration {
     }
 
 
+    /**
+     * @return pom.xml of this project.
+     */
     public File getPomFile() {
         String property = cfg.getProperty("pomFile");
         if (property == null) {
@@ -83,17 +116,58 @@ public class TckConfiguration {
     }
 
 
+    /**
+     * @return JDK home directory
+     */
     public File getJdkDirectory() {
         return new File(cfg.getProperty("jdk.directory"));
     }
 
 
+    /**
+     * @return Ant home directory (note: 1.9.14 recommended)
+     */
     public File getAntDirectory() {
         return new File(cfg.getProperty("ant.directory"));
     }
 
 
+    /**
+     * @return version of the GlassFish zip artifact.
+     */
     public String getGlassFishVersion() {
         return cfg.getProperty("glassfish.version");
+    }
+
+
+    /**
+     * @return logging.properties for the TCK's java client.
+     */
+    public File getClientLoggingProperties() {
+        return new File(cfg.getProperty("log.properties.client"));
+    }
+
+
+    /**
+     * @return logging.properties for both servers, ri and vi.
+     */
+    public File getServerLoggingProperties() {
+        return new File(cfg.getProperty("log.properties.server"));
+    }
+
+
+    /**
+     * @return true if you want to see verbose TCK's harness logs.
+     */
+    public boolean isHarnessLoggingEnabled() {
+        return Boolean.parseBoolean(cfg.getProperty("log.harness"));
+    }
+
+
+    /**
+     * @return true if you want to see logs of the communication of the asadmin wih the server.
+     */
+    public boolean isAsadminLoggingEnabled() {
+        return Boolean.parseBoolean(cfg.getProperty("log.asadmin"));
     }
 }
