+++ appserver/tests/tck/tck-runner/src/test/java/org/glassfish/main/tests/tck/ant/tests/AuthorizationITest.java
@@ -14,13 +14,14 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.main.tests.tck.ant;
+package org.glassfish.main.tests.tck.ant.tests;
 
 import jakarta.inject.Inject;
 
 import java.nio.file.Path;
 
-import org.junit.jupiter.api.Disabled;
+import org.glassfish.main.tests.tck.ant.TckRunner;
+import org.glassfish.main.tests.tck.ant.junit.TckTestExtension;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
@@ -28,34 +29,16 @@ import org.junit.jupiter.api.extension.ExtendWith;
  * @author David Matejcek
  */
 @ExtendWith(TckTestExtension.class)
-public class IntegrationITest {
+public class AuthorizationITest {
 
     @Inject
     private TckRunner tck;
 
     /**
-     * Usual time: 6 minutes
+     * 40 tests, usual time: 10 minutes.
      */
     @Test
-    public void integration() throws Exception {
-        tck.start(Path.of("integration"));
-    }
-
-    /**
-     * Usual time: 8 minutes
-     */
-    @Test
-    @Disabled("Requires mail server available; 6 failures.")
-    // TODO: Perhaps some mail server in test containers would be helpful.
-    public void samples() throws Exception {
-        tck.start(Path.of("samples"));
-    }
-
-    /**
-     * Usual time: 10 minutes
-     */
-    @Test
-    public void signaturetest() throws Exception {
-        tck.start(Path.of("signaturetest", "javaee"));
+    public void jacc() throws Exception {
+        tck.start(Path.of("jacc"));
     }
 }
