+++ appserver/tests/tck/tck-runner/src/test/java/org/glassfish/main/tests/tck/ant/tests/ExpressionLanguageITest.java
@@ -14,12 +14,14 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.main.tests.tck.ant;
+package org.glassfish.main.tests.tck.ant.tests;
 
 import jakarta.inject.Inject;
 
 import java.nio.file.Path;
 
+import org.glassfish.main.tests.tck.ant.TckRunner;
+import org.glassfish.main.tests.tck.ant.junit.TckTestExtension;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
@@ -27,25 +29,16 @@ import org.junit.jupiter.api.extension.ExtendWith;
  * @author David Matejcek
  */
 @ExtendWith(TckTestExtension.class)
-public class TransactionsITest {
+public class ExpressionLanguageITest {
 
     @Inject
     private TckRunner tck;
 
     /**
-     * Usual time: ? minutes
+     * Usual time: 32 minutes.
      */
     @Test
-    public void jta() throws Exception {
-        tck.start(Path.of("jta"));
-    }
-
-
-    /**
-     * Usual time: ? minutes
-     */
-    @Test
-    public void xa() throws Exception {
-        tck.start(Path.of("xa"));
+    public void el() throws Exception {
+        tck.start(Path.of("el"));
     }
 }
