+++ appserver/tests/tck/tck-runner/src/test/java/org/glassfish/main/tests/tck/ant/tests/PersistenceITest.java
@@ -14,12 +14,14 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.main.tests.tck.ant;
+package org.glassfish.main.tests.tck.ant.tests;
 
 import jakarta.inject.Inject;
 
 import java.nio.file.Path;
 
+import org.glassfish.main.tests.tck.ant.TckRunner;
+import org.glassfish.main.tests.tck.ant.junit.TckTestExtension;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
@@ -27,25 +29,43 @@ import org.junit.jupiter.api.extension.ExtendWith;
  * @author David Matejcek
  */
 @ExtendWith(TckTestExtension.class)
-public class WebservicesITest {
+public class PersistenceITest {
 
     @Inject
     private TckRunner tck;
 
-    /**
-     * Usual time: 90 minutes
-     */
     @Test
-    public void webservices12() throws Exception {
-        tck.start(Path.of("webservices12"));
+    public void appManaged() throws Exception {
+        tck.start(Path.of("jpa_appmanaged"));
     }
 
 
-    /**
-     * Usual time: 20 minutes
-     */
     @Test
-    public void webservices13() throws Exception {
-        tck.start(Path.of("webservices13"));
+    public void appManagedNoTx() throws Exception {
+        tck.start(Path.of("jpa_appmanagedNoTx"));
+    }
+
+
+    @Test
+    public void pmServlet() throws Exception {
+        tck.start(Path.of("jpa_pmservlet"));
+    }
+
+
+    @Test
+    public void puServlet() throws Exception {
+        tck.start(Path.of("jpa_puservlet"));
+    }
+
+
+    @Test
+    public void stateful3() throws Exception {
+        tck.start(Path.of("jpa_stateful3"));
+    }
+
+
+    @Test
+    public void stateless3() throws Exception {
+        tck.start(Path.of("jpa_stateless3"));
     }
 }
