+++ appserver/tests/tck/tck-runner/src/test/java/org/glassfish/main/tests/tck/ant/tests/ServerPagesITest.java
@@ -14,12 +14,14 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.main.tests.tck.ant;
+package org.glassfish.main.tests.tck.ant.tests;
 
 import jakarta.inject.Inject;
 
 import java.nio.file.Path;
 
+import org.glassfish.main.tests.tck.ant.TckRunner;
+import org.glassfish.main.tests.tck.ant.junit.TckTestExtension;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
@@ -27,31 +29,16 @@ import org.junit.jupiter.api.extension.ExtendWith;
  * @author David Matejcek
  */
 @ExtendWith(TckTestExtension.class)
-public class JdbcITest {
+public class ServerPagesITest {
 
     @Inject
     private TckRunner tck;
 
-    @Test
-    public void appClient() throws Exception {
-        tck.start(Path.of("jdbc_appclient"));
-    }
-
-
-    @Test
-    public void ejb() throws Exception {
-        tck.start(Path.of("jdbc_ejb"));
-    }
-
-
+    /**
+     * Usual time: ? minutes
+     */
     @Test
     public void jsp() throws Exception {
-        tck.start(Path.of("jdbc_jsp"));
-    }
-
-
-    @Test
-    public void servlet() throws Exception {
-        tck.start(Path.of("jdbc_servlet"));
+        tck.start(Path.of("jsp"));
     }
 }
