+++ appserver/tests/tck/tck-runner/src/test/resources/server-logging.properties
@@ -1,0 +1,23 @@
+
+handlers=org.glassfish.main.jul.handler.GlassFishLogHandler
+
+org.glassfish.main.jul.handler.GlassFishLogHandler.buffer.capacity=10000
+org.glassfish.main.jul.handler.GlassFishLogHandler.buffer.timeoutInSeconds=0
+org.glassfish.main.jul.handler.GlassFishLogHandler.enabled=true
+org.glassfish.main.jul.handler.GlassFishLogHandler.encoding=UTF-8
+org.glassfish.main.jul.handler.GlassFishLogHandler.file=${com.sun.aas.instanceRoot}/logs/server.log
+org.glassfish.main.jul.handler.GlassFishLogHandler.flushFrequency=1
+org.glassfish.main.jul.handler.GlassFishLogHandler.formatter=org.glassfish.main.jul.formatter.ODLLogFormatter
+org.glassfish.main.jul.handler.GlassFishLogHandler.formatter.excludedFields=
+org.glassfish.main.jul.handler.GlassFishLogHandler.formatter.multiline=true
+org.glassfish.main.jul.handler.GlassFishLogHandler.formatter.printSource=false
+org.glassfish.main.jul.handler.GlassFishLogHandler.level=ALL
+org.glassfish.main.jul.handler.GlassFishLogHandler.redirectStandardStreams=true
+org.glassfish.main.jul.handler.GlassFishLogHandler.rotation.compress=false
+org.glassfish.main.jul.handler.GlassFishLogHandler.rotation.maxArchiveFiles=0
+org.glassfish.main.jul.handler.GlassFishLogHandler.rotation.limit.megabytes=1024
+org.glassfish.main.jul.handler.GlassFishLogHandler.rotation.limit.minutes=0
+org.glassfish.main.jul.handler.GlassFishLogHandler.rotation.rollOnDateChange=false
+
+# See logging.properties in this repository to find other logger names.
+.level=INFO
