+++ appserver/tests/tck/tck-runner/src/test/resources/tck.properties
@@ -1,8 +1,13 @@
-glassfish.version=${project.version}
-tck.version=${project.version}
+glassfish.version=${glassfish.version}
+tck.version=${tck.version}
 
 target.directory=${project.build.directory}
 pomFile=${basedir}/pom.xml
 
-ant.directory=/home/dmatej/work/apache-ant-1.9.14
-jdk.directory=/usr/lib/jvm/jdk17
+ant.directory=${ant.home}
+jdk.directory=${jdk.home}
+
+log.harness=false
+log.asadmin=false
+log.properties.client=${project.build.testOutputDirectory}/client-logging.properties
+log.properties.server=${project.build.testOutputDirectory}/server-logging.properties
